phase3-ai-chatbot/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Phase II Full-Stack Web Todo Application - Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.32.1",
    "sqlmodel==0.0.22",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "pydantic[email]==2.12.5",
    "pydantic-settings==2.7.1",
    "bcrypt==4.2.1",
    "python-jose[cryptography]==3.3.0",
    "mcp==1.26.0",
    "openai==2.16.0",
    "openai-agents==0.7.0",
    "python-multipart==0.0.20",
    "passlib[bcrypt]==1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
